NMR-6982: remove user col, rename sf to frequency and restrict decimals, use same TableCell for regions and datasetbrowser
