+++ nmrfx-analyst-gui/src/main/java/org/nmrfx/analyst/gui/datasetbrowser/DatasetBrowserTableView.java
@@ -7,11 +7,13 @@ import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import org.controlsfx.control.tableview2.TableView2;
 import org.nmrfx.fxutil.Fx;
+import org.nmrfx.analyst.gui.utitlity.DoubleTableCell;
 import org.nmrfx.utilities.DatasetSummary;
 
 import java.util.List;
 
 public class DatasetBrowserTableView extends TableView2<DatasetSummary> {
+    private static final int NUMBER_DECIMAL_PLACES_FREQUENCY = 1;
     /* Keeps track of the summaries, new summaries are added to this list. */
     private final ObservableList<DatasetSummary> unfilteredDatasetSummaries = FXCollections.observableArrayList();
     private Runnable datasetSelectionListener = null;
@@ -22,9 +24,6 @@ public class DatasetBrowserTableView extends TableView2<DatasetSummary> {
 
         TableColumn<DatasetSummary, String> dateCol = new TableColumn<>("Date");
         dateCol.setCellValueFactory(new PropertyValueFactory<>("Time"));
-
-        TableColumn<DatasetSummary, String> userCol = new TableColumn<>("User");
-        userCol.setCellValueFactory(new PropertyValueFactory<>("User"));
 
         TableColumn<DatasetSummary, Boolean> presentCol = new TableColumn<>("InCache");
         presentCol.setCellValueFactory(new PropertyValueFactory<>("Present"));
@@ -39,10 +38,11 @@ public class DatasetBrowserTableView extends TableView2<DatasetSummary> {
         TableColumn<DatasetSummary, Integer> ndCol = new TableColumn<>("NDim");
         ndCol.setCellValueFactory(new PropertyValueFactory<>("nd"));
 
-        TableColumn<DatasetSummary, Double> sfCol = new TableColumn<>("SF");
+        TableColumn<DatasetSummary, Double> sfCol = new TableColumn<>("Frequency");
         sfCol.setCellValueFactory(new PropertyValueFactory<>("sf"));
+        sfCol.setCellFactory(column -> new DoubleTableCell<>(NUMBER_DECIMAL_PLACES_FREQUENCY));
 
-        getColumns().addAll(pathCol, userCol, dateCol);
+        getColumns().addAll(pathCol, dateCol);
         if (addCacheColumn) {
             getColumns().add(presentCol);
         }
