+++ nmrfx-analyst-gui/src/main/java/org/nmrfx/analyst/gui/utitlity/DoubleTableCell.java
@@ -1,0 +1,24 @@
+package org.nmrfx.analyst.gui.utitlity;
+
+import javafx.scene.control.TableCell;
+
+/**
+ * Table cell formatter to format non-editable columns of doubles
+ */
+public class DoubleTableCell<S, T extends Double> extends TableCell<S, Double> {
+    String formatString;
+
+    public DoubleTableCell(int decimalPlaces) {
+        formatString = "%." + decimalPlaces + "f";
+    }
+
+    @Override
+    protected void updateItem(Double value, boolean empty) {
+        super.updateItem(value, empty);
+        if (empty) {
+            setText(null);
+        } else {
+            setText(String.format(formatString, value));
+        }
+    }
+}
